# app/Modules/Asset/Requests/AssetStoreRequest.php
<?php

namespace App\Modules\Asset\Requests;

use Illuminate\Foundation\Http\FormRequest;
use Illuminate\Validation\Rule;

class AssetStoreRequest extends FormRequest
{
    public function rules(): array
    {
        return [
            'name' => ['required', 'string', 'min:3'],
            'type' => ['required', 'string', 'min:3'],
            'description' => ['string','min:3'],
        ];
    }
}