# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Booking TZ Api",
        "description": "Documentation for booking app",
        "contact": {
            "email": "some@mail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:88",
            "description": "API host"
        }
    ],
    "paths": {
        "/api/v1/resources": {
            "get": {
                "tags": [
                    "Booking resource"
                ],
                "summary": "Получить список всех ресурсов",
                "operationId": "594929af649719a6451ebf65e1dc9c12",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssetsResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/resources/": {
            "post": {
                "tags": [
                    "Booking resource"
                ],
                "summary": "Создать ресурс",
                "description": "",
                "operationId": "createAsset",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Asset"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/bookings/": {
            "post": {
                "tags": [
                    "Booking"
                ],
                "summary": "Создать бронирование",
                "description": "",
                "operationId": "createBooking",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BookingStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Booking"
                                }
                            }
                        }
                    },
                    "400, 422": {
                        "description": "Error"
                    }
                }
            }
        },
        "/api/v1/resources/{id}/bookings": {
            "get": {
                "tags": [
                    "Booking",
                    "Booking resource"
                ],
                "summary": "Получить всех бронирований для ресурса",
                "operationId": "assetBookingsShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the resource",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BookingsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error"
                    }
                }
            }
        },
        "/api/v1/bookings/{id}": {
            "delete": {
                "tags": [
                    "Booking"
                ],
                "summary": "Booking cancel",
                "operationId": "deleteBooking",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "123"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    },
                    "422": {
                        "description": "Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Asset": {
                "title": "Booking object",
                "required": [
                    "id",
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Conference Room A"
                    },
                    "type": {
                        "type": "string",
                        "example": "room"
                    },
                    "description": {
                        "type": "string",
                        "example": "Большая конференц-зал на 20 мест"
                    }
                },
                "type": "object"
            },
            "Booking": {
                "title": "Booking",
                "required": [
                    "asset_id",
                    "user_id",
                    "start_time",
                    "end_time"
                ],
                "properties": {
                    "asset_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-22T09:14:48.000000"
                    },
                    "end_time": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-02-22T09:14:48.000000"
                    }
                },
                "type": "object"
            },
            "AssetBookingsRequest": {
                "title": "AssetBookingsRequest",
                "description": "Asset Bookings",
                "properties": {
                    "id": {
                        "description": "asset id",
                        "type": "integer",
                        "example": "123"
                    }
                },
                "type": "object"
            },
            "AssetStoreRequest": {
                "title": "AssetStoreRequest",
                "description": "Description",
                "properties": {
                    "name": {
                        "description": "title",
                        "type": "string",
                        "example": "Asset title"
                    },
                    "type": {
                        "description": "type",
                        "type": "string",
                        "example": "Asset text"
                    },
                    "description": {
                        "description": "desc",
                        "type": "string",
                        "example": "Asset text"
                    }
                },
                "type": "object"
            },
            "AssetsResponse": {
                "title": "AssetResponse",
                "description": "Asset Response",
                "properties": {
                    "data": {
                        "description": "$data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Asset"
                        }
                    }
                },
                "type": "object"
            },
            "BookingStoreRequest": {
                "title": "BookingStoreRequest",
                "description": "Booking request",
                "properties": {
                    "asset_id": {
                        "description": "ID of the Asset",
                        "type": "integer",
                        "example": "123"
                    },
                    "user_id": {
                        "description": "ID of the User",
                        "type": "integer",
                        "example": "123"
                    },
                    "start_time": {
                        "description": "start booking time",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-08 19:08:50"
                    },
                    "end_time": {
                        "description": "end booking time",
                        "type": "string",
                        "format": "date",
                        "example": "2025-03-08 19:08:50"
                    }
                },
                "type": "object"
            },
            "BookingsResponse": {
                "title": "BookingsResponse",
                "description": "Asset Bookings response",
                "properties": {
                    "data": {
                        "description": "$data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Booking"
                        }
                    }
                },
                "type": "object"
            },
            "SuccessResponse": {
                "title": "SuccessResponse",
                "description": "Success",
                "properties": {
                    "data": {
                        "description": "$data",
                        "properties": {
                            "status": {
                                "description": "status",
                                "type": "boolean",
                                "example": "true"
                            },
                            "errors": {
                                "description": "errors",
                                "type": "null",
                                "example": "null"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "400": {
                "description": "Bad request"
            },
            "400, 422": {
                "description": "Error"
            }
        }
    },
    "tags": [
        {
            "name": "Booking resource",
            "description": "Booking resource"
        },
        {
            "name": "Booking",
            "description": "Booking"
        }
    ]
}